Find out the return from always buying one unit
This serves as a baseline for calculating performance.  What would the
rate of return be if we would have just purchased one unit of the
SP500 every day over the last 5 years, and never sold?  This is
essentially a mock of the retirement plan.  We aren't buying to sell,
we are buying to hold.

We would have made a "profit" of 21%, at a cost of ~$3 million.